+++ build_runner_core/lib/src/generate/build_runner.dart
@@ -19,8 +19,9 @@ class BuildRunner {
 
   Future<Null> beforeExit() => _build.beforeExit();
 
-  Future<BuildResult> run(Map<AssetId, ChangeType> updates) =>
-      _build.run(updates);
+  Future<BuildResult> run(Map<AssetId, ChangeType> updates,
+          {List<String> buildDirs}) =>
+      _build.run(updates, buildDirs: buildDirs);
 
   static Future<BuildRunner> create(
           BuildOptions options,