+++ build_runner_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- The `build` method now requires a list of `buildDirs`.
+- Remove `buildDirs` from `BuildOptions`.
+
 ## 1.1.3
 
 - Update to `package:graphs` version `0.2.0`.